pyproject.toml
[project]
name = "voice-lab"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[silero,openai]",
    "websockets>=12.0",
    "pyyaml",
]

[project.optional-dependencies]
raw = [
    "httpx",
    "numpy",
    "pyaudio",
]
server-stt = [
    "fastapi",
    "uvicorn[standard]",
    "faster-whisper",
    "numpy",
]
server-tts = [
    "fastapi",
    "uvicorn[standard]",
    "piper-tts",
]

[tool.setuptools.packages.find]
include = ["pipeline*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
